msct_shape: honor remove_temp_files option (#1624)
Former-commit-id: 765dd661b2deabe58fb91b29ffa9a320a826f10b [formerly 77213fdaa3ef51684f26deb1aeeed487d8a2f8c5 [formerly 2fd864b98e99342560c66449d3aaeb7c496b1f30]]
Former-commit-id: 9c2c4db43fdadd422de26a5cf4421efb6ed7587a
Former-commit-id: e1d09626cf1c51e7f02b90752a32fcb029e217cb

+++ scripts/msct_shape.py
@@ -367,7 +367,8 @@ def compute_properties_along_centerline(fname_seg_image, property_list, fname_di
 
     # Removing temporary folder
     os.chdir(curdir)
-    sct.rmtree(path_tmp)
+    if remove_temp_files:
+        sct.rmtree(path_tmp)
 
     return property_list, properties
